koszxyk

+++ src/js/koszyk.js
@@ -1,21 +1,30 @@
 function addToKoszyk(produkt_id) {
     // Get items
-    let cart = JSON.parse(localStorage.getItem('koszyk')) || [];
+    let cart = getKoszyk();
 
     // Add item to cart
     cart.push(produkt_id);
 
     // Save data to localStorage
     localStorage.setItem('koszyk', JSON.stringify(cart));
-
-    return cart;
 };
 
-function pokazKoszyk() {
+function removeFromKoszyk(produkt_id) {
+    let cart = getKoszyk()
+
+    cart = cart.filter(id => id !== produkt_id);
+
+    localStorage.setItem('koszyk', JSON.stringify(cart));
+}
+
+function getKoszyk() {
     const cart = JSON.parse(localStorage.getItem('koszyk')) || [];
     return cart;
 };
 
 function clearKoszyk() {
-    localStorage.clear();
+    let user = confirm("Are you siure to delete entire koszyk??");
+    if (user) {
+        localStorage.clear();
+    }
 }
